src/data_mining_twitter/requirements.txt
oauthlib==0.7.2
python-twitter==2.2
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
tweepy==3.3.0
twython==3.2.0
